fix: password strength is now in tsx
